export to MIDI and Lilypond
This adds a "list" mode to the app.

List mode allows you to replay marked sections.
Rename/retitle marked sections.
Delete marked sections.
Playback marked sections.
Export to MIDI
Export to Lilypond

Is it now possible to smash on the keyboard for a bit, work out a ditty and
export it to Lilypond to make something out of it? Why, yes, it kind of is.

Fixes #4. Export to Lilypond