better Lilypond export
We investigated abjad, but it wasn't a good fit for our use-case. However, the
ability to tweak the note-length at a later date was integrated in to our
solution.

It no longer produces half-assed lilypond files that need major work to get the
chords to line up with the melody. The new format has some infrastructure in
place to support lyrics, too.

closes #14