refactoring + import with sample playback
Major changes.

We have a shared "genericlister" that is used by both the jam lister as well as the import selector and can also be leveraged for the spreadsheet-based import process.

We support non-blocking curses.getkey operations within the same thread, also supporting better timeouts than normally supported by curses (using fractional seconds) and not needing to perform a catch if a time-out occurs. (It returns None, which makes sense.)

closes #23