refactoring + import with sample playback   (#24)
* major improvements to "list" mode
* major refactoring overall
* import file-selector
* preview with seek for import file-selector

We have a shared "genericlister" that is used by both the jam lister as well as the import selector and can also be leveraged for the spreadsheet-based import process. It also supports large lists now, and assigning keys to commands.

I thought we'd be playing with WAV files, but my target WAV files are apparently incompatible. (24 bit instead of 16 appears to cause a problem.) Fortunately, MP3 are just as useful at this stage. In the long-term, if the backend supports MP3, Ogg Vorbis and FLAC, I should be good.

We support non-blocking curses.getkey operations within the same thread, also supporting better timeouts than normally supported by curses (using fractional seconds) and not needing to perform a catch if a time-out occurs. (It returns None, which makes sense.)

The import file-selector works with MP3s (only, right now) and allows previewing the file, as well as seeking by minutes and seconds using the `<`/`>` and `[`/`]` keys. (Regardless of shift state.)

closes #23