Initial pass of the import screen
You can use "p" to play from the current position to the end of the file.

You can use the "." key to mark the current moment as a beat, so we can now listen to a file and type the beats out.

"d" deletes the row.

Right now, Enter/Return repeats the current beat, where a beat starts at the current file location and goes to the next file location.

The "@" key jumps to a specific location, creating that location if it doesn't already exist.

closes #20