add lyrics and songs to import mode
And just like that, we have a concept of lyrics and songs.

While the songs _exist_, there isn't much to the notion right now. No time, tempo, key or scale just yet. Still, it is something to hang those off of. We hang the title off of the notion of the "mark" for the note.

(Which reminds me, this adds support for setting the "mark". Now you can indicate verses, chorus, etc.)

The lyrics, on the other hand... They're not exactly what I _wanted_, but they'll work for the FAWM-release.

closes #26
closes #23