first pass at integrating chords
It's a bit wobbly in places, but seems remarkably operational.

This also adds the ability to shift note metadata up or down. Since we want to minimize keys, and we know we don't want to go beyond the next note marker, we do a simple half-up or half-down based on the key pressed. This allows a person to perform a binary search by hand to find the point they're looking for.

closes #29
closes #28