import mode supports export to plain-text
We're kind of wiggling around the various export formats for the import mode. Right now the export to Lilypond and MIDI processes have commands in the UI, but are broken.

Export to text, however, works quite nicely.

This starts the use of "\" for a stanza break and "/" for a line break, following on the use of those characters in the KAR format. This works for the text-export, and there's a command to toggle that setting without editing the lyric, going to the beginning of the line and changing it. (It's the `/` command.)

The csvplayer has some concept of exporting, but appears to only properly export the lyrics and metadata. (So a little more work to be done there.)

The `genericlister` and the `spreader` based upon it, now use their own `FieldReader` class to read input, allowing us to perform our normal idle call. The up and down keys jump by word, which comes in handy for some lyric use-cases. The new class also allows us to update existing data, instead of just requiring a whole new line to be entered. Lyrics and Marks support this.

closes #37