Heading in to polish... maybe?
This adds the ability to clone rows instead of just moving rows. We still use the same VI-like keys, but the upper-case version clones the row instead of moving it.

This adds a weird way of specifying the lead key. `[` and `]` go to the previous/next key in the scale. `{` and `}` go to the previous/next key chromatically. `'` goes up an octave. `,` goes down an octave.

This fixes all kinds of export-from-import-mode issues. Lilypond export now includes lyrics, lead notes and pad chords. The Lilypond mode looks really good these days. The MIDI export works again.

The `h` key plays just the MIDI events without the audio. This includes both the pad as well as lead.

The `;` key will briefly play the lead note.

closes #39
closes #35
closes #32