tab when entering lyrics does next-in-kind
It turned out it was trivial to implement next-in-kind for any command that takes a string. (Mostly lyrics and marks.)

What this means:

You're entering lyrics for a note. This note is done and you're ready to move to the next. Hit TAB and it accepts that line, moves you down one, and starts editing the lyrics for the line below it.

This ran in to potential issues when playing back just that one line. Now we're more clear that when playing back just one line, you're not playing back _that_ line, you're always playing back the _active_ line, so as you go up and down the list you'll hear it jump to the current line.

This makes it so that when you are listening to just that line while typing lyrics, and you hit TAB, you'll start hearing the next line.

closes #46