minor fixes
Track "0.0" is used for the initial data -- if present. (Instead of always starting with -1 from the "no track here" logic.)

There were some cases where an empty string for a string-based command would kill the app.

Nudge doesn't work at the top/bottom, so it presents a message about this.

Nudging plays a trick on the current line to cause the "play current line" logic to refresh. This makes it _much_ nicer.

The default song chord progression wasn't being set properly, so when it was changed by the user, it was ignored.

We have an "apply-change" command when getting a string. (^R) This is to accompany the "next-in-kind" (TAB) command. While "next-in-kind" does the same command on the line below it, "apply-change" redoes the command on the current line. (Not sure it will be super-useful after I was playing with it, but it works.)

Here's something new: chord progressions can fiddle with the octave using the standard "'" and "," characters as suffixes. This means we can tweak just the octave in an otherwise standard Nashville chord progression.