more tweaks
Don't touch the file if we don't do anything. This adds an explicit "changed" state. This seemed important as we now sort the import list by timestamp (most recent on top), and we didn't want to touch a file when we just wanted to listen to it.

The app no longer defaults to jam mode. It'll produce help text if you don't specify the mode.

Lilypond export has been tweaked to deal with trailing " --" and " __" which would be inserted on the end of a note's lyric. Previously it was (wrongly) quoting it.

Text export has been cleaned up. "Next Screen" markers leave an empty line, but can't accidentally produce multiple empty lines. A trailing untracked entry no longer messes up the last line of text. Tracks are separated with two empty lines.

The Import Lister now supports export of MIDI, Lilypond, and plain-text of the items in the list. It also uses the basename for the default filename, instead of defaulting to "export". Since there weren't a lot of commands defined for the import lister, we've bound both the standard Importer's commands, as well as the more logical initial letter.