some important fixes (#52)
When audio files were over an hour long, it messed with the importer display. That's been fixed.

Key_Up without lyrics now won't kill the app anymore.

The Import Lister now supports "D" to delete the audio file. It doesn't actually delete it, but there's no in-app method to recover it. (You need to manually tweak the meta file.) This helps when dealing with a lot of audio files, and some of them have been fully processed.