various important fixes
Using the "bpm" key to assign key slots now accepts multipliers for the BPM, so you can say you want half notes for 142 BPM instead of quarter notes.

We now have explicit commands to swap line contents with the line below ("w") and the line above ("W") when in import mode.

Since the "lily_key" isn't derived from the standard "key", we now use the same default for the "key" that we use elsewhere.

Speaking of which, we now default to G Major instead of C Major.

All of the lister-type UIs now have page-up and page-down using shift-up and shift-down. We should probably actually support the page-up and page-down keys at some point, but Macs don't have them and that's what I'm working on.

We don't currently export the drum/metronome track in the jamlister, so we now no longer play it, too.

Drum velocity is significantly lowered, as it is just a metronome for now. The background player had some major refactory, as it wasn't wanting to stop/start previously.

On a related note, the pad volume for the jam-mode was reset to normal levels.

There was a problem with the preset not being specified in jam file when it was triggered from a mark-based event. It turned out, we just weren't writing it. Now we do.