bug(lilypond export): lilypond now produces expected results
It's now possible to take an MP3 of silence in to the import tool, mark it up according to 100 BPM, give it notes and chords and lyrics and produce correct sheet music on the other side.

This means while the fuzzy logic may still be a bit fuzzy, the basic logic is sound.

closes #60