integrate aubio and improve chord mode
Aubio-based automation is available via the "a" key. It works best when the tracks are marked, but the lyrics are not.

There's now a concept of "chord selctions", and if it is in selection-mode, any configured chord is available at any time. (As opposed to just going to the next chord.) This is available as pad_selections, and uses the same names as the normal chord configuration.

This also changes the UI for the chords, so that they more clearly show the configured chord.

It also fixes "jam" (currently called "gui") mode and "panic" mode.

closes #71
closes #66