integrate aubio and improve chord mode (#72)
* add beat-repeat with the "B" key

Is there an established rhythm for a section of the file? Leverage it.

12 x 120bpm -- create 12 1-beat locations at 120bpm
6 x 2@120bpm -- create 6 2-beat locations at 120bpm
6 beats -- between row and row+1 is a beat, repeat that 6 times.

* integrate aubio and improve chord mode

Aubio-based automation is available via the "a" key. It works best when the tracks are marked, but the lyrics are not.

There's now a concept of "chord selctions", and if it is in selection-mode, any configured chord is available at any time. (As opposed to just going to the next chord.) This is available as pad_selections, and uses the same names as the normal chord configuration.

This also changes the UI for the chords, so that they more clearly show the configured chord.

It also fixes "jam" (currently called "gui") mode and "panic" mode.

closes #71
closes #66